RAG性能调优/pyproject.toml
[project]
name = "learningrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "addict>=2.4.0",
    "asyncio>=4.0.0",
    "datasets>=4.8.4",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.3",
    "hnswlib>=0.8.0",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "llama-index>=0.14.19",
    "llama-index-core>=0.14.19",
    "llama-index-embeddings-dashscope>=0.5.0",
    "llama-index-embeddings-openai>=0.6.0",
    "llama-index-llms-dashscope>=0.6.0",
    "llama-index-llms-openai>=0.7.4",
    "llama-index-llms-openai-like>=0.7.1",
    "llama-index-retrievers-bm25>=0.7.1",
    "llama-index-tools-tavily-research>=0.5.0",
    "llama-index-vector-stores-faiss>=0.6.0",
    "llama-index-vector-stores-milvus>=1.1.0",
    "matplotlib>=3.10.8",
    "memory-profiler>=0.61.0",
    "nest-asyncio>=1.6.0",
    "openai>=2.30.0",
    "paddleocr>=3.4.0",
    "paddlepaddle==3.3.0",
    "pandas<3.0.0",
    "pillow>=12.1.1",
    "pip>=26.0.1",
    "pymilvus>=2.6.11",
    "python-dotenv>=1.2.2",
    "ragas>=0.4.3",
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.56.0",
    "tiktoken>=0.12.0",
    "timm>=1.0.26",
    "torch>=2.10.0",
    "transformers>=5.4.0",
    "unstructured[md]>=0.22.6",
    "uvicorn>=0.43.0",
]
